refactor: usePlayerPosition 훅 추가 및 StagOne 수정
